split app into many files due to growing project size
