+++ app.py
@@ -1,49 +1,16 @@
-# standard libs
 from contextlib import asynccontextmanager
-from datetime import datetime, timedelta
+from datetime import datetime
 import json
-import re
-import time
-import uuid
-# third party libs
-from apscheduler.schedulers.background import BackgroundScheduler
-from apscheduler.triggers.cron import CronTrigger
-from fastapi import FastAPI, HTTPException, Request, Response
-from fastapi import status as HTTPStatusCode
-from pydantic import BaseModel
-from nacl.signing import VerifyKey
-from nacl.exceptions import BadSignatureError
-import pytz
-import redis
-import requests
-from tinydb import TinyDB, where
-# custom code
-import env
-import ref
+from fastapi import FastAPI, Response, status as HTTPStatusCode
+from db import REDIS, REDIS_KEY, maps_table, subs_table
+import routers.admin, routers.interaction
 
 
 
-DISCORD_PUBLIC_KEY = '59a9c5881d2c0f19456a150b2d9b2b8b203e568164614a2815f7e505c62faa50'
-DISCORD_VERIFIER = VerifyKey(bytes.fromhex(DISCORD_PUBLIC_KEY))
-
-REDIS = redis.Redis(host=env.REDIS_HOST, port=env.REDIS_PORT, username=env.REDIS_USERNAME, password=env.REDIS_PASSWORD, decode_responses=True)
-REDIS_KEY = f"bot.{env.ENV_NAME}.db"
-
-DISCORD_HEADERS = {'Content-Type':'application/json', 'Authorization':f"Bot {env.DISCORD_BOT_TOKEN}"}
-FETCH_HEADERS = {'User-Agent' : f"kcjwv-icy-cotd-bot-{env.ENV_NAME}"}
-CONTROL_CHARS_PATTERN = '\\$(?:[wnoitsgz]|[0-9A-F]{3})'
-
-SCHEDULER = BackgroundScheduler(timezone='CET')
-CET_TZ = pytz.timezone('CET')
-
-
-
-db = TinyDB(REDIS_KEY)
-maps_table = db.table('maps')
-subs_table = db.table('subscriptions')
-
+# lifecycle manager to handle startup/shutdown tasks
 @asynccontextmanager
 async def lifespan(app: FastAPI):
+
     # startup - load db from redis or local file
     print('Loading DB backup')
     maps = REDIS.get(f"{REDIS_KEY}:maps")
@@ -54,8 +21,10 @@ async def lifespan(app: FastAPI):
     if subs:
         subs_table.insert_multiple(json.loads(subs))
         print('Loaded subscriptions successfully')
+
     # yield - let application run
     yield
+
     # cleanup - export db file back to redis
     print('Exporting DB backup')
     maps = maps_table.all()
@@ -69,307 +38,13 @@ async def lifespan(app: FastAPI):
 
 
 app = FastAPI(lifespan=lifespan)
+app.include_router(routers.admin.router)
+app.include_router(routers.interaction.router)
 
-
-
-# global app settings that can be updated by the /settings route
-class Settings:
-    notifications_enabled = env.NOTIFICATIONS_ENABLED_DEFAULT
-
-
-
-"""
-GET '/'
-Simple healthcheck endpoint.
-"""
 @app.get('/')
 def root():
-    return {
+    content = {
         'current_time' : datetime.utcnow().isoformat(),
         'subscriptions_count' : len(subs_table.all()),
     }
-
-
-
-"""
-POST '/interaction'
-Endpoint where Discord interactions will be sent.
-https://discord.com/developers/docs/interactions/receiving-and-responding#receiving-an-interaction
-"""
-class InteractionType:
-    PING = 1
-    CHAT = 4
-
-class Command:
-    HELP = 'help'
-    def help():
-        pass
-
-
-    SHOW = 'show'
-    def show():
-        pass
-    
-
-    SUBSCRIBE = 'subscribe'
-    def subscribe(guild_id:int, channel_id:int, role_id:int, style:str) -> str:
-        subscription_id = str(uuid.uuid4())
-        j = {
-            'subscription_id' : subscription_id,
-            'guild_id' : guild_id,
-            'channel_id' : channel_id,
-            'role_id' : role_id,
-            'style' : style.lower(),
-        }
-        # enforce uniqueness for combination guild_id + style
-        subs_table.remove((where('guild_id') == j['guild_id']) & (where('style') == j['style']))
-        subs_table.insert(j)
-        return subscription_id
-
-
-    STYLES = 'styles'
-    def styles():
-        results = list(ref.TMX_MAP_TAGS.values())
-        results.sort()
-        return results
-    
-
-    UNSUBSCRIBE = 'unsubscribe'
-    def unsubscribe(guild_id:int, style:str) -> int:
-        j = {
-            'guild_id' : guild_id,
-            'style' : style.lower(),
-        }
-        removed = subs_table.remove((where('guild_id') == j['guild_id']) & (where('style') == j['style']))
-        return len(removed)
-
-
-
-@app.post('/interaction')
-async def interaction(req:Request):
-    raw_body = await req.body()
-    body = raw_body.decode()
-    signature = req.headers.get('X-Signature-Ed25519')
-    timestamp = req.headers.get('X-Signature-Timestamp')
-
-    # respond to discord's security tests
-    if env.VERIFY_SIGNATURES:
-        try:
-            DISCORD_VERIFIER.verify(f"{timestamp}{body}".encode(), bytes.fromhex(signature))
-        except BadSignatureError:
-            raise HTTPException(status_code=HTTPStatusCode.HTTP_401_UNAUTHORIZED, detail='Invalid request signature')
-    j = await req.json()
-    if j['type'] == InteractionType.PING:
-        content = {'type':InteractionType.PING}
-        return Response(content=json.dumps(content), status_code=HTTPStatusCode.HTTP_200_OK, media_type='application/json')
-    
-    # set up default message
-    message = 'Oops - something went wrong'
-
-    # handle slash commands
-    guild_id = j['guild_id']
-    channel_id = j['channel_id']
-    command = j['data']['name']
-    options = {}
-    if j['data'].get('options'):
-        for option in j['data']['options']:
-            name, value = option['name'], option['value']
-            options[name] = value
-
-    if command == Command.HELP:
-        help_text = Command.help()
-        message = "Your command worked, but it hasn't been implemented yet"
-
-    elif command == Command.SHOW:
-        subscriptions = Command.show()
-        message = "Your command worked, but it hasn't been implemented yet"
-    
-    elif command == Command.SUBSCRIBE:
-        role_id, style = options.get('role'), options.get('style')
-        if not role_id or not style:
-            raise HTTPException(status_code=HTTPStatusCode.HTTP_422_UNPROCESSABLE_ENTITY, detail='Role and Style are required')
-        subscription_id = Command.subscribe(guild_id, channel_id, role_id, style)
-        print(f"Created subscription {subscription_id} for server {guild_id}")
-        message = f"Successfully subscribed to {style.upper()} - notifications will be posted to this channel"
-
-    elif command == Command.STYLES:
-        styles_list = Command.styles()
-        styles_fmt = '\n'.join([f"{i+1}. {s}" for i,s in enumerate(styles_list)])
-        message = f"All of the following are valid style names:\n{styles_fmt}"
-
-    elif command == Command.UNSUBSCRIBE:
-        style = options.get('style')
-        if not style:
-            raise HTTPException(status_code=HTTPStatusCode.HTTP_422_UNPROCESSABLE_ENTITY, detail='Style is required')
-        num_removed = Command.unsubscribe(guild_id, style)
-        print(f"Removed {num_removed} subscriptions for server {guild_id}")
-        if num_removed == 0:
-            message = f"Subscription not found for {style.upper()} - nothing to delete"
-        else:
-            message = f"Unsubscribed from {style.upper()}"
-
-    # format into discord json and return
-    content = {
-        'type': InteractionType.CHAT,
-        'data': {
-            'content' : message,
-        }
-    }
-    return Response(content=json.dumps(content), status_code=HTTPStatusCode.HTTP_200_OK, media_type='application/json')
-
-
-
-"""
-NOTIFY JOB
-Scheduled job, invoked by the end of REFRESH JOB like a DAG. Sends style notifications to applicable Discord channels.
-"""
-def notify_job():
-    # query maps table
-    today = datetime.strftime(datetime.now(CET_TZ), '%Y-%m-%d')
-    yesterday = datetime.strftime(datetime.now(CET_TZ) + timedelta(days=-1), '%Y-%m-%d')
-    map_record = maps_table.get(where('date') == today)
-    if not map_record:
-        map_record = maps_table.get(where('date') == yesterday)
-    tags_lower = [t.lower() for t in map_record['tags']]
-    # query subscriptions table
-    subscriptions = subs_table.search(where('style').map(lambda s:s.lower()).one_of(tags_lower))
-    print(f"Found {len(subscriptions)} matching subscriptions")
-    # build message payloads
-    messages = []
-    for sub_record in subscriptions:
-        message = f"<@&{sub_record['role_id']}>\n{sub_record['style'].upper()} COTD: \"{map_record['map_name']}\" by {map_record['author_name']} (AT: {map_record['author_time']})"
-        subscription_id = sub_record['subscription_id']
-        target_url = f"https://discord.com/api/channels/{sub_record['channel_id']}/messages"
-        t = (subscription_id, target_url, message)
-        messages.append(t)
-    # push to all subscribers
-    print(f"Pushing {len(messages)} notifications")
-    for sub_id, url, msg in messages:
-        body = json.dumps({'content' : msg})
-        resp = requests.post(url, data=body, headers=DISCORD_HEADERS)
-        print(f"{sub_id}: {resp.status_code}")
-
-
-
-"""
-REFRESH JOB
-Scheduled job to retrieve the new COTD at 7pm CET / 1pm EST every day.
-"""
-def refresh_job():
-    refresh_timestamp = datetime.utcnow().isoformat()
-
-    # retrieve map from trackmania.io
-    url = 'https://trackmania.io/api/totd/0'
-    resp = requests.get(url, headers=FETCH_HEADERS)
-    print(f"trackmania.io: {resp.status_code}")
-    tmio_json = resp.json()
-    map_uid = tmio_json['days'][-1]['map']['mapUid']
-    print(f"totd map_uid: {map_uid}")
-
-    # retrieve same map from tmx
-    url = f"https://trackmania.exchange/api/maps/get_map_info/uid/{map_uid}"
-    resp = requests.get(url, headers=FETCH_HEADERS)
-    print(f"tmx: {resp.status_code}")
-    tmx_json = resp.json()
-
-    # extract all useful information
-    tags = [int(t) for t in tmx_json['Tags'].split(',')]
-    tags_readable = [ref.TMX_MAP_TAGS[t] for t in tags]
-    y,m,d = tmio_json['year'], tmio_json['month'], len(tmio_json['days'])
-    map_date = f"{y}-{m}-{d}"
-    map_json = tmio_json['days'][-1]['map']
-    map_name_cleaned = re.sub(CONTROL_CHARS_PATTERN, '', map_json['name'])
-    author_name = map_json['authorplayer']['name']
-    author_time = map_json['authorScore'] / 1000
-
-    # write to db, enforcing uniqueness by date
-    output = {
-        'date' : map_date,
-        'map_name' : map_name_cleaned,
-        'author_name' : author_name,
-        'author_time' : author_time, 
-        'tags' : tags_readable,
-        'updated' : refresh_timestamp,
-    }
-    print(output)
-    maps_table.remove(where('date') == output['date'])
-    maps_table.insert(output)
-    if Settings.notifications_enabled:
-        print('Triggering notifications')
-        time.sleep(1)
-        SCHEDULER.add_job(notify_job, next_run_time=datetime.now(CET_TZ))
-    else:
-        print('Notifications disabled - done')
-
-
-
-# admin routes below
-
-
-class AdminBody(BaseModel):
-    admin_key: str
-
-
-"""
-GET/PUT '/settings'
-Admin route where global settings can be modified.
-For example, enable/disable notifications to prevent accidental @Role spam on restart.
-"""
-@app.get('/settings')
-def get_settings():
-    settings = {k:v for k,v in Settings.__dict__.items() if not callable(v) and not str(k).startswith('__')}
-    return Response(json.dumps(settings), status_code=HTTPStatusCode.HTTP_200_OK)
-
-class SettingsBody(AdminBody):
-    notifications_enabled:bool | None = None
-
-@app.put('/settings')
-def update_settings(body:SettingsBody):
-    if body.admin_key != env.ADMIN_KEY:
-        raise HTTPException(status_code=HTTPStatusCode.HTTP_401_UNAUTHORIZED, detail='Invalid admin key')
-    if body.notifications_enabled:
-        Settings.notifications_enabled = body.notifications_enabled
-    return Response(status_code=HTTPStatusCode.HTTP_200_OK)
-
-
-
-"""
-POST '/refresh'
-Admin route to trigger data refresh & notifications, if enabled.
-"""
-class RefreshBody(AdminBody):
-    pass
-
-@app.post('/refresh')
-def refresh(body:RefreshBody):
-    if body.admin_key != env.ADMIN_KEY:
-        raise HTTPException(status_code=HTTPStatusCode.HTTP_401_UNAUTHORIZED, detail='Invalid admin key')
-    SCHEDULER.add_job(refresh_job, next_run_time=datetime.now(CET_TZ))
-    content = {'message':'Data refresh triggered'}
-    return Response(content=json.dumps(content), status_code=HTTPStatusCode.HTTP_202_ACCEPTED)
-
-
-
-"""
-POST '/reset'
-Admin route to hard reset all subscriptions. Disallowed in production.
-"""
-class ResetBody(AdminBody):
-    pass
-
-@app.post('/reset')
-def reset(body:ResetBody):
-    if body.admin_key != env.ADMIN_KEY:
-        raise HTTPException(status_code=HTTPStatusCode.HTTP_401_UNAUTHORIZED, detail='Invalid admin key')
-    if 'prod' in env.ENV_NAME.lower():
-        raise HTTPException(status_code=HTTPStatusCode.HTTP_403_FORBIDDEN, detail='Reset is disallowed in production')
-    REDIS.delete(REDIS_KEY)
-    with open(REDIS_KEY, 'w') as f:
-        json.dump(dict(), f)
-    return Response(status_code=HTTPStatusCode.HTTP_204_NO_CONTENT)
-
-
-
-# kick off background tasks
-SCHEDULER.add_job(refresh_job, CronTrigger.from_crontab('0 19 * * *'), retry_on_exception=True)
-SCHEDULER.start()
+    return Response(content=json.dumps(content), status_code=HTTPStatusCode.HTTP_200_OK)
